blog_project/blogproject/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
whitenoise==6.6.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7